@mgkuhn 2020-11-25
Thanks for pointing out this huge inefficiency! However, is always first allocating the `samples` array in `UInt64` and then bulk-converting it in the end to `sample_type` really the best we can do here? Shouldn't `sample_type` really be a parametric type of the function, such that the compiler can then specialize all of `read_pcm_samples` for each value of `sample_type`?

I think how this function deals with types needs a slightly bigger rethink.