@mgkuhn 2020-11-25
I see that
```
function read_ieee_float_samples(io::IO, chunk_size,
                                 fmt::WAVFormat, subrange,
                                 ::Type{floatType}) where {floatType}
```
does already use a type parameter `floatType`, and I suspect `read_pcm_samples` should do something similar. The call to `pcm_container_type(nbits)` needs to move out of `read_pcm_samples` such that the compiler can dispatch on the result.